Data Type Fix

+++ database/migrations/2022_03_12_141150_update_shows_table.php
@@ -15,7 +15,7 @@ class UpdateShowsTable extends Migration
     {
         Schema::table('shows', function (Blueprint $table) {
             $table->date('premiered')->nullable();
-            $table->tinyInteger('thetvdb')->unsigned();
+            $table->integer('thetvdb')->unsigned();
         });
 
         Schema::table('episodes', function (Blueprint $table) {
